code/Modified-Previous-Code/requirements.txt
carla==0.9.13
keras==2.7.0
numpy==1.21.5
opencv-python==4.5.4.60
pandas==1.1.5
Pillow==8.4.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
PyYAML==6.0
matplotlib==3.5.1
seaborn>=0.11.2
moviepy
